+++ guillotina/tests/image/test_field.py
@@ -1,7 +1,10 @@
 from guillotina.contrib.image.behaviors import IImageAttachment
 from guillotina.contrib.image.behaviors import IMultiImageAttachment
 from guillotina.contrib.image.behaviors import IMultiImageOrderedAttachment
+from guillotina.utils import get_behavior
 from guillotina.tests.image import TEST_DATA_LOCATION
+from guillotina.directives import index_field
+from guillotina.test_package import IExample
 
 import json
 import os
@@ -9,6 +12,18 @@ import pytest
 
 
 pytestmark = pytest.mark.asyncio
+
+NOT_POSTGRES = os.environ.get("DATABASE", "DUMMY") in ("cockroachdb", "DUMMY")
+PG_CATALOG_SETTINGS = {
+    "applications": ["guillotina", "guillotina.contrib.catalog.pg", "guillotina.contrib.image"],
+    "cloud_datamanager": "db",
+    "load_utilities": {
+        "catalog": {
+            "provides": "guillotina.interfaces.ICatalogUtility",
+            "factory": "guillotina.contrib.catalog.pg.utility.PGSearchUtility",
+        }
+    },
+}
 
 
 @pytest.mark.app_settings(
@@ -111,7 +126,7 @@ async def test_multiimage_field_with_behavior(redis_container, container_request
             "POST",
             "/db/guillotina/",
             data=json.dumps(
-                {"@type": "Item", "@behaviors": [IMultiImageAttachment.__identifier__], "id": "foobar"}
+                {"@type": "Example", "@behaviors": [IMultiImageAttachment.__identifier__], "id": "foobar"}
             ),
         )
         assert status == 201
@@ -138,6 +153,33 @@ async def test_multiimage_field_with_behavior(redis_container, container_request
         assert status == 200
 
 
+@index_field.with_accessor(IExample, "images")
+async def index_images(obj):
+    behavior = await get_behavior(obj, IMultiImageOrderedAttachment)
+    result = []
+    if behavior.images:
+        for key in behavior.images.keys():
+            result.append(key)
+    return result
+
+
+@index_field.with_accessor(
+    IExample,
+    "images",
+    behavior="guillotina.contrib.image.behaviors.IMultiImageOrderedAttachment",
+    type="object",
+)
+async def index_images_behavior(obj):
+    behavior = await get_behavior(obj, IMultiImageOrderedAttachment)
+    result = []
+    if behavior.images:
+        for key in behavior.images.keys():
+            result.append(key)
+    return result
+
+
+@pytest.mark.app_settings(PG_CATALOG_SETTINGS)
+@pytest.mark.skipif(NOT_POSTGRES, reason="Only PG")
 @pytest.mark.app_settings(
     {"applications": ["guillotina", "guillotina.contrib.image"], "cloud_datamanager": "db"}
 )
@@ -145,12 +187,15 @@ async def test_multiimage_ordered_field_with_behavior(redis_container, container
     async with container_requester as requester:
         _, status = await requester("POST", "/db/guillotina/@addons", data=json.dumps({"id": "image"}))
         assert status == 200
-
         response, status = await requester(
             "POST",
             "/db/guillotina/",
             data=json.dumps(
-                {"@type": "Item", "@behaviors": [IMultiImageOrderedAttachment.__identifier__], "id": "foobar"}
+                {
+                    "@type": "Example",
+                    "@behaviors": [IMultiImageOrderedAttachment.__identifier__],
+                    "id": "foobar",
+                }
             ),
         )
         assert status == 201
@@ -209,6 +254,10 @@ async def test_multiimage_ordered_field_with_behavior(redis_container, container
         response, status = await requester("DELETE", "/db/guillotina/foobar/@delete/images/key1")
         assert status == 200
 
+        response, status = await requester("GET", "/db/guillotina/@search?id=foobar")
+        assert status == 200
+        assert response["items"][0]["images"] == ["key2", "key0", "key3"]
+
         response, status = await requester("GET", "/db/guillotina/foobar")
         behavior = response["guillotina.contrib.image.behaviors.IMultiImageOrderedAttachment"]
         keys_ordered = {"key2": 0, "key0": 1, "key3": 2}
@@ -216,6 +265,9 @@ async def test_multiimage_ordered_field_with_behavior(redis_container, container
         for image in behavior["images"].keys():
             assert count == keys_ordered[image]
             count += 1
+
+        response, status = await requester("GET", "/db/guillotina/@search?id=foobar")
+        assert status == 200
 
         response, status = await requester(
             "PATCH", "/db/guillotina/foobar/@sort/images", data=json.dumps(["key3", "key2", "key0"])
@@ -229,3 +281,6 @@ async def test_multiimage_ordered_field_with_behavior(redis_container, container
         for image in behavior["images"].keys():
             assert count == keys_ordered[image]
             count += 1
+        response, status = await requester("GET", "/db/guillotina/@search?id=foobar")
+        assert status == 200
+        assert response["items"][0]["images"] == ["key3", "key2", "key0"]